+++ src/main/java/com/projet/equipement/dto/ligneAchat/LigneAchatUpdateDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class LigneAchatUpdateDto {
 
-    private Integer prixAchatUnitaire;
+    private Double prixAchatUnitaire;
 
     private Integer quantite;
